+++ CHANGES.rst
@@ -10,6 +10,7 @@
 - Removed :meth:`pyudev.Enumerator.match_children` in favour of
   :meth:`pyudev.Enumerator.match_parent`
 - :attr:`pyudev.Device.children` requires udev version 172 now
+- #31: Added :meth:`pyudev.Enumerator.match_attribute`
 
 
 0.12 (Aug 31, 2011)